// BEEQT/src/app/core/models/user.ts
export class User {
  id: string;
  email: string;
  password: string;
  userName: string;
  phoneNumber: string;
  isAdmin: boolean;
  avatar: string;
  createdOn: string;
  firstName: string;
  lastName: string;
  address1: string;
  address2: string;
  country: string;
  state: string;
}


export class Account {
  id: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  avatar: string;
  address1: string;
  country: string;
  state: string;
}